Set dossier filename
